requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
web3==6.11.0
flask==2.3.3
gunicorn==21.2.0
apscheduler==3.10.4
cryptography==41.0.7
aiohttp==3.9.1